requirements.txt
fastapi>=0.103.0
uvicorn[standard]>=0.22.0
websockets>=11.0
python-multipart>=0.0.6
pandas>=1.5.0
scikit-learn>=1.3.0
python-dotenv>=1.0.0
httpx>=0.25.0
psycopg2-binary>=2.9.9
gTTS>=2.5.1
google-generativeai>=0.8.0
pytest>=7.4.0
pytest-asyncio>=0.21.0
